internode: multi-step-correctness (M1+M2) — 1M+ training steps safe
Retrofits the cumulative-ring-control protocols from C2-C4 with two
narrow fixes that close the failure modes that would only surface
across many training iters. Smoke-green at production shape (6.5
ms/iter, ~0.4 ms within the C4 baseline's per-allocation variance).
**Preserves all of C4's cleanup elision — no team-sync, no IBGDA
quiet, no NVL barriers, no inter-iter memsets re-introduced.** Plan
captured in `multi-step-correctness.md`.

M1 — C4 force-write at iter entry
─────────────────────────────────
NVL gen-stamp's 12-bit seq window (4096 iters) is safe for
adjacent-iter aliasing, but two scenarios slip through:

(a) **Sparse-routing dormancy.** A `(channel, dst_nvl)` substream
    with zero token flow for ≥ 4096 consecutive iters then resumes
    flow — the slot's stale value carries an aliased seq tag and the
    forwarder/receiver consumes it as current. Possible with hot/cold
    expert imbalance.
(b) **Same-iter cross-phase aliasing.** Fwd dispatch and bwd
    dispatch_grads share the same `tile_signal.dispatch_seq` within
    a layer (see `stream_moe.py:509, 543, 584, 615`), so bwd's first
    read of a slot would naively seq-match fwd's leftover value.
    Today saved only by "bwd writes every slot fwd wrote before any
    reader gets there" — fragile.

Fix: in `dispatch_main_kernel` and `dispatch_grads_main_kernel`, the
forwarder publishes `nvl_pack(seq, 0)` to `nvl_channel_tail.buffer()`
and the NVL receiver publishes the same to `nvl_channel_head.buffer()`
at kernel entry, unconditionally — one store per substream per iter
via `elect_one_sync()`. The reader's gen-stamp spin protocol is
unchanged; force-write just guarantees that by the time any reader
looks, the slot carries the current iter's seq tag. No barrier
added — the receiver still gates on prefix_start/end seq-match
(written later by the forwarder's meta-poll branch) before
inspecting tail, and the forwarder's backpressure check tolerates
stale-tagged reads via the seq-match filter.

Cost: 4 stores × 32 channels × 8 NVL peers = ~1K stores/iter,
parallelized across SMs, sub-µs total.

M2 — 8-byte NIC AMO + int64 reader_prev arrays
──────────────────────────────────────────────
C2's head/tail counters use mlx5 4-byte `MLX5_OPCODE_ATOMIC_MASKED_FA`
with `IBGDA_4_BYTE_EXT_AMO_OPMOD`; the slot is `SymBuffer<uint64_t>`
but only the low 4 bytes are atomically incremented, and `reader_prev_*`
arrays are int32. Wrap horizon at production routing (~128
tokens/iter/slot, 2 phases/layer) is ~250k steps × 32 layers — within
realistic training duration. Once `cur` wraps to negative, signed
`atomicMax` keeps the old positive `prev` forever and `cur - prev`
math degrades into multi-iter accumulation → forwarder timeouts.

Fix:

1. New WQE constructor `ibgda_write_amo_add_wqe_8byte` in
   `ibgda_device.cuh`: opcode `MLX5_OPCODE_ATOMIC_FA` (no MASKED, no
   4-byte opmod), `mlx5_wqe_atomic_seg` with `swap_add` set,
   `compare = 0`, `data_seg.byte_count = sizeof(int64_t)`. New wrapper
   `nvshmemi_ibgda_amo_nonfetch_add_8byte` for both RDMA and local-
   copy (8-byte `atomicAdd` on `unsigned long long*`).
2. Switch 6 head/tail AMO sites in `internode.cu` (3 dispatch +
   3 dispatch_grads + 2 combine) from the 4-byte helper to the
   8-byte helper.
3. Widen `dispatch_reader_prev_head/tail`, `combine_reader_prev_head/tail`
   from `int*` to `int64_t*` in `Buffer` (deep_ep.hpp), allocate at
   `sizeof(int64_t)` in `Buffer::sync`. Plumb int64 through
   `DispatchEnv::reader_prev_*` and the `launch_combine_main` /
   `combine_main_kernel` signatures.
4. All 6 read-subtract sites (forwarder/sender/receiver seeds) widen
   the math to int64 to avoid the inner `static_cast<int>` truncating
   the cumulative counter to its low 32 bits before subtraction:
   `static_cast<int>(static_cast<int64_t>(ld_*(...)) - prev_at_entry)`.
5. All 6 atomicMax writeback sites reinterpret the int64 pointer as
   `unsigned long long*` (CUDA atomicMax's universal overload). The
   counter is unsigned-monotonic so signed/unsigned semantics
   coincide.

C3 sentinel stays int32: only increments by 1 per iter per
(channel, dst_rdma); wrap horizon 2^31 ≈ 2.1B = ~30+ years of typical
training. Not worth widening.

Wrap horizon at int64: 2^63 ≈ 9.2 × 10^18. Effectively infinite.

Smoke. `srun_internode.sh -m stream_ep.stream_moe.smoke_pipeline
--num_sms 64`: green, 6.5 ms/iter at world=16, T=8192. Within
per-allocation Slurm variance of the 6.1 ms C4 baseline. Stress run
deferred to M3.

Next: M3 — synthetic skewed-routing stress harness (10k+ iters with
hot/cold experts) to validate M1's force-write fix against the
sparse-routing aliasing path that smoke can't surface in 5 iters.

+++ csrc/deep_ep.hpp
@@ -195,10 +195,13 @@ private:
     // Sized [max_num_channels × num_rdma_ranks] int32, ~512 B each at
     // production; ~2 KB total. Allocated zero-init in `Buffer::Buffer`,
     // freed in destructor.
-    int* dispatch_reader_prev_head = nullptr;
-    int* dispatch_reader_prev_tail = nullptr;
-    int* combine_reader_prev_head  = nullptr;
-    int* combine_reader_prev_tail  = nullptr;
+    // int64_t (M2): the NIC AMOs that drive these counters were widened from
+    // 4-byte to 8-byte (mlx5 MLX5_OPCODE_ATOMIC_FA); int32 wraps at ~250k
+    // training steps × 32 layers, int64 horizon is effectively infinite.
+    int64_t* dispatch_reader_prev_head = nullptr;
+    int64_t* dispatch_reader_prev_tail = nullptr;
+    int64_t* combine_reader_prev_head  = nullptr;
+    int64_t* combine_reader_prev_tail  = nullptr;
     // Persistent prev-sentinel array for the RDMA dispatch meta region (C3).
     // The meta SymBuffer's slot 30 ("kRdmaMetaSentinelSlot") accumulates
     // across iters via amo_nonfetch_add — sender bulk_puts the 18 data ints
